Move acata & friends into Data.Higher.Functor.Foldable.
