+++ src/Control/Higher/Comonad/Cofree.hs
@@ -1,10 +1,8 @@
 {-# LANGUAGE FlexibleContexts, InstanceSigs, RankNTypes, PolyKinds, ScopedTypeVariables, TypeFamilies, TypeOperators #-}
 module Control.Higher.Comonad.Cofree where
 
-import Control.Arrow
 import Data.Higher.Bifunctor
 import Data.Higher.Functor
-import Data.Higher.Functor.Foldable
 import Data.Higher.Product
 import Data.Higher.Transformation
 
@@ -22,21 +20,6 @@ unwrap :: Cofree f a ~> f (Cofree f a)
 unwrap = tailF . runCofree
 
 
-acata :: forall c t. (HFunctor (Base t), Recursive t) => (Base t c ~> c) -> t ~> Cofree (Base t) c
-acata f = go
-  where go :: t ~> Cofree (Base t) c
-        go = cofree . uncurry (:<) . (f . hfmap extract &&& id) . hfmap go . project
-
-apara :: forall c t. (HFunctor (Base t), Recursive t) => (Base t (t :*: c) ~> c) -> t ~> Cofree (Base t) c
-apara f = go
-  where go :: t ~> Cofree (Base t) c
-        go = cofree . uncurry (:<) . (f . hfmap (hsecond extract) &&& hfmap hsnd) . hfmap ((:*:) <*> go) . project
-
-
-unannotate :: (HFunctor (Base t), Corecursive t) => Cofree (Base t) c ~> t
-unannotate = cata (embed . tailF)
-
-
 -- Instances
 
 instance HFunctor f => HBifunctor (CofreeF f) where
@@ -50,9 +33,3 @@ instance HFunctor f => HFunctor (Cofree f) where
   hfmap f = go
     where go :: Cofree f a ~> Cofree f b
           go = cofree . hbimap f go . runCofree
-
-
-type instance Base (Cofree f v) = CofreeF f v
-
-instance HFunctor f => Recursive (Cofree f v) where project = runCofree
-instance HFunctor f => Corecursive (Cofree f v) where embed = cofree
