+++ src/Data/Higher/Functor/Foldable.hs
@@ -1,6 +1,9 @@
 {-# LANGUAGE FlexibleContexts, PolyKinds, RankNTypes, ScopedTypeVariables, TypeFamilies, TypeOperators #-}
 module Data.Higher.Functor.Foldable where
 
+import Control.Arrow ((&&&))
+import Control.Higher.Comonad.Cofree
+import Data.Higher.Bifunctor
 import Data.Higher.Functor
 import Data.Higher.Product
 import Data.Higher.Sum
@@ -44,8 +47,29 @@ hylo f g = go
 newtype Fix f a = Fix { unFix :: f (Fix f) a }
 
 
+acata :: forall c t. (HFunctor (Base t), Recursive t) => (Base t c ~> c) -> t ~> Cofree (Base t) c
+acata f = go
+  where go :: t ~> Cofree (Base t) c
+        go = cofree . uncurry (:<) . (f . hfmap extract &&& id) . hfmap go . project
+
+apara :: forall c t. (HFunctor (Base t), Recursive t) => (Base t (t :*: c) ~> c) -> t ~> Cofree (Base t) c
+apara f = go
+  where go :: t ~> Cofree (Base t) c
+        go = cofree . uncurry (:<) . (f . hfmap (hsecond extract) &&& hfmap hsnd) . hfmap ((:*:) <*> go) . project
+
+
+unannotate :: (HFunctor (Base t), Corecursive t) => Cofree (Base t) c ~> t
+unannotate = cata (embed . tailF)
+
+
 -- Instances
 
 type instance Base (Fix f) = f
 
 instance HFunctor f => Recursive (Fix f) where project = unFix
+
+
+type instance Base (Cofree f v) = CofreeF f v
+
+instance HFunctor f => Recursive (Cofree f v) where project = runCofree
+instance HFunctor f => Corecursive (Cofree f v) where embed = cofree
